key部分をbucketに統合
